pyproject.toml
[tool.poetry]
name = "geoname-sandbox"
version = "0.1.0"
description = ""
authors = ["Stefane Fermigier <sf@abilian.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

devtools = ">=0.12.2"
attrs = ">=23.2.0"
advanced-alchemy = "^0.9.0"
tqdm = "^4.66.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
